backend/pyproject.toml

[tool.poetry]
name = "rag-ui-backend"
version = "0.1.0"
description = "RAG UI Backend with FastAPI"
authors = ["Gemini <gemini@google.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
langchain = "^0.2.6"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.5"
requests = "^2.32.3"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.7"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
